Update QRCode.php
Until PHP8 it seems it was possible to use an array for the colors in the qrcode-config and hand them over to the qrcode. As of now, the type of the parameter is important and has to be correct. My "quick-n-dirty" fix is to get the arrays für foreground and background and override the corresponding options (line 36-42).

Then we surely need to use $this->options in the foreach-loop since $options would reference the unchanged version. This change should be used definitely since $options is written in $this->options and we should avoid making mistakes by using two variables that could contain different arrays after line 29.

With these changes the plugin works for my configuration with PHP 8.0.8.